refactor: reorganize course components, add NoCoursesFound and CareerPathVisualizer, and move CourseUtils to ui directory
