+++ components/cursos/NoCoursesFound.tsx
@@ -1,26 +1,31 @@
+
+/**
+ * A React functional component that displays a message indicating that no courses were found.
+ * This component is styled with Tailwind CSS and includes an icon, a title, and a description.
+ *
+ * @component
+ * @returns {JSX.Element} A styled message box indicating no courses were found.
+ *
+ * @example
+ * // Usage in a parent component
+ * import NoCoursesFound from './NoCoursesFound';
+ *
+ * function App() {
+ *   return (
+ *     <div>
+ *       <NoCoursesFound />
+ *     </div>
+ *   );
+ * }
+ */
+import { Award } from 'lucide-react';
+
 export default function NoCoursesFound() {
-    return (
-      <div className="bg-white rounded-lg shadow-sm p-12 text-center">
-        <div className="w-16 h-16 bg-gray-100 rounded-full flex items-center justify-center mx-auto mb-4">
-          <svg 
-            xmlns="http://www.w3.org/2000/svg" 
-            className="h-8 w-8 text-gray-400" 
-            fill="none" 
-            viewBox="0 0 24 24" 
-            stroke="currentColor"
-          >
-            <path 
-              strokeLinecap="round" 
-              strokeLinejoin="round" 
-              strokeWidth={1.5} 
-              d="M9 12h6m-6 4h6m2 5H7a2 2 0 01-2-2V5a2 2 0 012-2h5.586a1 1 0 01.707.293l5.414 5.414a1 1 0 01.293.707V19a2 2 0 01-2 2z" 
-            />
-          </svg>
-        </div>
-        <h2 className="text-xl font-medium mb-2">No se encontraron cursos</h2>
-        <p className="text-gray-500 mb-6">
-          No tienes cursos que cumplan con los filtros seleccionados.
-        </p>
-      </div>
-    );
-  }
+  return (
+    <div className=" rounded-lg p-12 text-center bg-gray-50 rounded-lg p-8">
+      <Award className="h-12 w-12 text-gray-400 mx-auto mb-3" />
+      <h3 className="text-lg font-medium text-gray-500 mb-1">No hay certificaciones</h3>
+      <p className="text-gray-500">No se encontraron certificaciones con los filtros actuales.</p>
+    </div>
+  );
+}
