+++ components/ui/CourseUtils.tsx
